species now using taxa as base and aggregating associated name data. updated species photo endpoints to remove index interface

+++ core/migrations/2023-09-25-090522_create_name_data_summaries/up.sql
@@ -8,7 +8,7 @@ SELECT
     COALESCE(markers.total, 0) + COALESCE(genomes.total, 0) + COALESCE(other_data.total, 0) AS total_genomic
 FROM names
 LEFT JOIN (
-     SELECT name_id, count(*) AS total
+     SELECT name_id, count(*)::int AS total
      FROM sequencing_events
      JOIN sequences ON sequencing_events.sequence_id = sequences.id
      WHERE target_gene IS NOT NULL
@@ -16,20 +16,20 @@ LEFT JOIN (
 ) markers ON markers.name_id = names.id
 
 LEFT JOIN (
-     SELECT name_id, count(*) AS total
+     SELECT name_id, count(*)::int AS total
      FROM assembly_events
      JOIN sequences ON assembly_events.sequence_id = sequences.id
      GROUP BY name_id
 ) genomes ON genomes.name_id = names.id
 
 LEFT JOIN (
-     SELECT name_id, count(*) AS total
+     SELECT name_id, count(*)::int AS total
      FROM specimens
      GROUP BY name_id
 ) specimens ON specimens.name_id = names.id
 
 LEFT JOIN (
-     SELECT name_id, count(*) as total
+     SELECT name_id, count(*)::int as total
      FROM sequences
      LEFT JOIN sequencing_events se on sequences.id = se.sequence_id
      LEFT JOIN annotation_events ae on sequences.id = ae.sequence_id
